requirements.txt
# Core Dependencies
streamlit==1.32.0
python-dotenv==1.0.0
sqlalchemy==2.0.28
bcrypt==4.1.2
stripe==7.14.0
passlib==1.7.4
watchdog==3.0.0
streamlit-option-menu==0.3.6


# Database Drivers (SQLite included by default)
sqlalchemy-utils==0.41.1

# Additional Utilities
pillow==10.2.0  # For image handling
python-multipart==0.0.9  # For file uploads